tests: split checker coverage by responsibility
