+++ src/check-test-harness.ts
@@ -1,0 +1,8 @@
+import { checkDesign } from "./check";
+
+export function messages(root: string): string[] {
+	return checkDesign(root).map(
+		(diagnostic) =>
+			`${diagnostic.path}:${diagnostic.line}:${diagnostic.column} TS${diagnostic.code}: ${diagnostic.message}`,
+	);
+}
